댓글 등록하기 오류 잡기 완료

+++ src/main/java/shop/mtcoding/blog/controller/BoardController.java
@@ -75,12 +75,25 @@ public class BoardController {
     }
 
     @PostMapping("/board/{id}/delete")
-    public String delete(@PathVariable Integer id) {
+    public String delete(@PathVariable Integer id) { // 1. PathVariable 값 받기
+        // 2.인증검사
+        // session에 접근해서 sessionUser 키값을 가져오세요
+        // null 이면, 로그인페이지로 보내고
+        // null 아니면, 3번을 실행하세요.
         User sessionUser = (User) session.getAttribute("sessionUser");
         if (sessionUser == null) {
-            return "redirect:/loginForm";
+            return "redirect:/loginForm"; // 401
         }
 
+        // 3. 권한검사
+        Board board = boardRepository.findById(id);
+        if (board.getUser().getId() != sessionUser.getId()) {
+            return "redirect:/40x"; // 403 권한없음
+        }
+
+        // 4. 모델에 접근해서 삭제
+        // boardRepository.deleteById(id); 호출하세요 -> 리턴을 받지 마세요
+        // delete from board_tb where id = :id
         boardRepository.deleteById(id);
 
         return "redirect:/";
@@ -160,11 +173,11 @@ public class BoardController {
         return "board/detail";// V
     }
 
-    @ResponseBody
-    @GetMapping("/test/reply")
-    public List<Reply> test2() {
-        List<Reply> replys = replyRepository.findByBoardId(1);
-        return replys;
-    }
+    // @ResponseBody
+    // @GetMapping("/test/reply")
+    // public List<Reply> test2() {
+    // List<Reply> replys = replyRepository.findByBoardId(1);
+    // return replys;
+    // }
 
 }
